@@ -1,72 +1,93 @@
 <?php
+// src/Controller/AuthController.php
 namespace App\Controller;
 
 use App\Entity\User;
-use App\Form\LoginType;
 use App\Form\RegisterType;
 use App\Repository\UserRepository;
+use Doctrine\ORM\EntityManagerInterface;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\PasswordHasher\Hasher\UserPasswordHasherInterface;
-use Symfony\Component\Routing\Attribute\Route;
+use Symfony\Component\Routing\Annotation\Route;
 use Symfony\Component\Security\Http\Authentication\AuthenticationUtils;
 
 class AuthController extends AbstractController
 {
-    #[Route('/register', name: 'app_register_page')]
-    public function index(Request $req, UserRepository $userRepository, UserPasswordHasherInterface $passwordHasher): Response
-    {
-        // Logique d'inscription
-        $user = new User();
-        $inscriptionForm = $this->createForm(RegisterType::class, $user);
-        $inscriptionForm->handleRequest($req);
+    // Route pour l'inscription
+    #[Route('/inscription', name: 'app_register')]
+    public function register(
+        Request $request,
+        UserRepository $userRepository, 
+        EntityManagerInterface $entityManager, 
+        UserPasswordHasherInterface $passwordHasher
+    ): Response {
+        $user = new User(); // Crée un nouvel utilisateur
 
-        if ($inscriptionForm->isSubmitted() && $inscriptionForm->isValid()) {
+        // Crée le formulaire d'inscription basé sur la classe RegisterType
+        $form = $this->createForm(RegisterType::class, $user);
+
+        // Gère la requête HTTP et lie les données au formulaire
+        $form->handleRequest($request);
+
+        // Si le formulaire est soumis et valide
+        if ($form->isSubmitted() && $form->isValid()) {
+
             $userExistant = $userRepository->findOneBy(['email' => $user->getEmail()]);
-            if ($userExistant) {
-                $this->addFlash('error', 'Utilisateur déjà inscrit !');
-                return $this->redirectToRoute('app_register_page');
-            }
 
+            if ($userExistant) {
+                // Ajouter un message flash et rediriger
+                $this->addFlash('error', 'Cet utilisateur existe déjà.');
+                return $this->redirectToRoute('app_register');
+            }
+            // Encode le mot de passe en utilisant UserPasswordHasherInterface
             $hashedPassword = $passwordHasher->hashPassword($user, $user->getPassword());
             $user->setPassword($hashedPassword);
-            $userRepository->save($user, true);
 
+            // Sauvegarde l'utilisateur en base de données
+            $entityManager->persist($user);
+            $entityManager->flush();
+
+            // Ajoute un message de succès
             $this->addFlash('success', 'Inscription réussie !');
-            return $this->redirectToRoute('app_login');
+
+            // Redirige vers la page d'accueil ou la page de connexion
+            return $this->redirectToRoute('app_home');
         }
 
-        // Formulaire de connexion
-        $connexionForm = $this->createForm(LoginType::class);
+        // Si la méthode est GET ou si le formulaire n'est pas valide, on rend la vue
         return $this->render('pages/auth/register.html.twig', [
-            'inscriptionForm' => $inscriptionForm->createView(),
-            'connexionForm' => $connexionForm->createView(),
+            'form' => $form->createView(), // Passe la vue du formulaire au template
         ]);
+
     }
 
-    #[Route('/login', name: 'app_login')]
-    public function connexionTraitement(AuthenticationUtils $authenticationUtils, Request $request): Response
-    {
-        // Récupérer l'erreur et l'email du dernier essai
-        $error = $authenticationUtils->getLastAuthenticationError();
-        $dernierEmail = $authenticationUtils->getLastUsername();
-
-        // Gérer la redirection après une connexion réussie
-        $targetPath = $request->get('target_path', $this->generateUrl('homepage')); // Par défaut redirige vers la page d'accueil
-
-        return $this->render('pages/auth/login.html.twig', [
-            'connexion' => [
-                'error' => $error ? $error->getMessageKey() : null,
-                'dernierEmail' => $dernierEmail,
-                'target_path' => $targetPath
-            ]
-        ]);
+    #[Route('/connexion', name: 'app_connexion')]
+public function login(AuthenticationUtils $authenticationUtils): Response
+{
+    // Vérifiez si l'utilisateur est déjà connecté
+    if ($this->getUser()) {
+        return $this->redirectToRoute('app_home');
     }
 
-    #[Route('/logout', name: 'app_logout')]
+    // Obtenez l'erreur d'authentification, s'il y en a
+    $error = $authenticationUtils->getLastAuthenticationError();
+    $lastUsername = $authenticationUtils->getLastUsername();
+
+    // Renvoyez la vue avec les erreurs et le dernier nom d'utilisateur saisi
+    return $this->render('pages/auth/login.html.twig', [
+        'last_username' => $lastUsername,
+        'error' => $error,
+    ]);
+}
+
+
+    // Route pour la déconnexion (facultatif, géré automatiquement par Symfony)
+    #[Route('/deconnexion', name: 'app_deconnexion')]
     public function logout(): void
     {
-        // Cette méthode peut rester vide car Symfony gère la déconnexion automatiquement
+        // Symfony gère déjà la déconnexion, cette méthode peut être vide
     }
+
 }
