@@ -10,19 +10,28 @@ use Symfony\Component\Form\FormBuilderInterface;
 
 class LoginType extends AbstractType
 {
-    public function buildForm(FormBuilderInterface $builder, array $options)
+    public function buildForm(FormBuilderInterface $builder, array $options): void
     {
         $builder
             ->add('email', EmailType::class, [
                 'label' => 'E-mail',
-                'attr' => ['placeholder' => 'Votre e-mail']
+                'attr' => [
+                    'placeholder' => 'Votre adresse e-mail',
+                    'class' => 'form-control', // Ajout d'une classe CSS pour le style
+                ],
             ])
             ->add('password', PasswordType::class, [
                 'label' => 'Mot de passe',
-                'attr' => ['placeholder' => 'Votre mot de passe']
+                'attr' => [
+                    'placeholder' => 'Votre mot de passe',
+                    'class' => 'form-control', // Ajout d'une classe CSS pour le style
+                ],
             ])
             ->add('submit', SubmitType::class, [
-                'label' => 'Se connecter'
+                'label' => 'Se connecter',
+                'attr' => [
+                    'class' => 'btn btn-primary w-100 mt-3', // Style bouton Bootstrap ou autre
+                ],
             ]);
     }
 }
