@@ -1,52 +1,59 @@
 <?php
-
-
 // src/Form/RegisterType.php
 namespace App\Form;
 
+use App\Entity\User;
 use Symfony\Component\Form\AbstractType;
-use Symfony\Component\Form\Extension\Core\Type\TextType;
 use Symfony\Component\Form\Extension\Core\Type\EmailType;
 use Symfony\Component\Form\Extension\Core\Type\PasswordType;
-use Symfony\Component\Form\Extension\Core\Type\TelType;
+use Symfony\Component\Form\Extension\Core\Type\TextType;
 use Symfony\Component\Form\Extension\Core\Type\SubmitType;
 use Symfony\Component\Form\FormBuilderInterface;
 use Symfony\Component\OptionsResolver\OptionsResolver;
 
 class RegisterType extends AbstractType
 {
-    public function buildForm(FormBuilderInterface $builder, array $options)
+    public function buildForm(FormBuilderInterface $builder, array $options): void
     {
         $builder
             ->add('firstName', TextType::class, [
                 'label' => 'Prénom',
-                'attr' => ['placeholder' => 'Votre prénom']
+                'attr' => [
+                    'placeholder' => 'Votre prénom',
+                ],
             ])
-            ->add('lastName', TextType::class, [
-                'label' => 'Nom',
-                'attr' => ['placeholder' => 'Votre nom']
+            ->add('pseudo', TextType::class, [
+                'label' => 'Pseudo',
+                'attr' => [
+                    'placeholder' => 'Votre pseudo',
+                ],
             ])
             ->add('email', EmailType::class, [
                 'label' => 'E-mail',
-                'attr' => ['placeholder' => 'Votre e-mail']
-            ])
-            ->add('phone', TelType::class, [
-                'label' => 'Téléphone',
-                'attr' => ['placeholder' => 'Votre téléphone']
+                'attr' => [
+                    'placeholder' => 'Votre adresse e-mail',
+                ],
             ])
             ->add('password', PasswordType::class, [
                 'label' => 'Mot de passe',
-                'attr' => ['placeholder' => 'Votre mot de passe']
+                'attr' => [
+                    'placeholder' => 'Votre mot de passe',
+                ],
             ])
             ->add('submit', SubmitType::class, [
-                'label' => 'S’inscrire'
+                'label' => 'S’inscrire',
+                'attr' => [
+                    'class' => 'btn btn-primary',
+                ],
             ]);
     }
 
-    public function configureOptions(OptionsResolver $resolver)
+    public function configureOptions(OptionsResolver $resolver): void
     {
         $resolver->setDefaults([
-            'data_class' => null, // Vous pouvez lier à un utilisateur si vous avez une entité User
+            'data_class' => User::class, // Lien vers l'entité User
         ]);
     }
 }
+
+
